Backendfeature/community

See merge request swm-13-main/13_swm17/snack-styling!24

+++ src/backend/spring/src/main/java/com/snackstyling/spring/domain/Answer.java
@@ -22,7 +22,7 @@ public class Answer {
     @JoinColumn(name="queId")
     private Question question;
     private LocalDateTime postDate;
-    private Integer codi;
+    private Long codi;
     @Column(length = 1000)
     private String comments;
 }
